+++ tests/unit/fitness/test_repair.py
@@ -25,7 +25,8 @@ def my_runresult(my_software):
 ])
 def test_process_inherit(my_software, my_runresult, return_code, status):
     exec_result = ExecResult(['(empty)'], 'SUCCESS', return_code, b'', b'', 1, 0)
-    my_software.fitness[0].process_init_exec(my_runresult, exec_result)
+    klass = magpie.utils.convert.fitness_from_string('repair')
+    klass(my_software).process_init_exec(my_runresult, exec_result)
     assert my_runresult.status == status, my_runresult
 
 @pytest.mark.parametrize(('stdout', 'status', 'fitness'), [